+++ src/test/scala/com/localytics/sbt/PidUtilsTest.scala
@@ -1,14 +1,13 @@
-package com.localytics.sbt.dynamodb
+package com.localytics.sbt
 
 import java.io.File
 
-import org.scalatest.{FunSpec, Matchers}
+import org.scalatest.FunSpec
+import org.scalatest.Matchers
 
-import scala.io.Source
+class PidUtilsTest extends FunSpec with Matchers {
 
-class DynamoDBLocalUtilsTest extends FunSpec with Matchers {
-
-  describe("DynamoDBLocalUtils") {
+  describe("PidUtils") {
 
     it("should extract PID correctly") {
       val jpsOutput =
@@ -20,8 +19,8 @@ class DynamoDBLocalUtilsTest extends FunSpec with Matchers {
           |92431 /Users/person/code/repository/aws-mocks/dynamo-db/DynamoDBLocal.jar -port 8000 -inMemory -sharedDb
           |74414 /usr/local/Cellar/sbt/0.13.7/libexec/sbt-launch.jar
         """.stripMargin
-      val baseDir = new File("/Users/person/code/repository/aws-mocks/dynamo-db/")
-      DynamoDBLocalUtils.extractDynamoDBPid(jpsOutput, 8000, baseDir) should equal(Some("92431"))
+      val jar = new File("/Users/person/code/repository/aws-mocks/dynamo-db/DynamoDBLocal.jar")
+      PidUtils.extractPid(jpsOutput, 8000, jar) should equal(Some("92431"))
     }
 
     it("should resolve multiple local dynamodb instances") {
@@ -34,26 +33,11 @@ class DynamoDBLocalUtilsTest extends FunSpec with Matchers {
           |92431 /Users/person/code/repository/aws-mocks/dynamo-db/DynamoDBLocal.jar -port 8000 -inMemory -sharedDb
           |74414 /usr/local/Cellar/sbt/0.13.7/libexec/sbt-launch.jar
         """.stripMargin
-      val baseDir = new File("/Users/person/code/repository/aws-mocks/dynamo-db/")
-      DynamoDBLocalUtils.extractDynamoDBPid(jpsOutput, 8000, baseDir) should equal(Some("92431"))
+      val jar = new File("/Users/person/code/repository/aws-mocks/dynamo-db/DynamoDBLocal.jar")
+      PidUtils.extractPid(jpsOutput, 8000, jar) should equal(Some("92431"))
 
     }
 
-    it("should identify a valid jar") {
-      DynamoDBLocalUtils.validJar(new File(getClass.getResource("/valid.jar").getFile)) should be(true)
-    }
-
-    it("should identify an invalid jar") {
-      DynamoDBLocalUtils.validJar(new File(getClass.getResource("/invalid.jar").getFile)) should be(false)
-    }
-
-    it("should identify a valid gz") {
-      DynamoDBLocalUtils.validGzip(new File(getClass.getResource("/valid.tar.gz").getFile)) should be(true)
-    }
-
-    it("should identify an invalid gz") {
-      DynamoDBLocalUtils.validGzip(new File(getClass.getResource("/invalid.tar.gz").getFile)) should be(false)
-    }
   }
 
 }
